mod: DIで実際のDBを使うように変更

+++ app/src/main/java/com/ogata_k/mobile/winp/presentation/event/work/FailedCreateWork.kt
@@ -2,4 +2,4 @@ package com.ogata_k.mobile.winp.presentation.event.work
 
 import com.ogata_k.mobile.winp.presentation.event.Event
 
-data class FailedCreateWork(val workId: Int) : Event
+data class FailedCreateWork(val workId: Long) : Event
