+++ app/src/main/java/com/ogata_k/mobile/winp/presentation/event/work/SucceededUpdateWork.kt
@@ -2,4 +2,4 @@ package com.ogata_k.mobile.winp.presentation.event.work
 
 import com.ogata_k.mobile.winp.presentation.event.Event
 
-data class SucceededUpdateWork(val workId: Int) : Event
+data class SucceededUpdateWork(val workId: Long) : Event
