+++ app/src/main/java/com/ogata_k/mobile/winp/presentation/event/work/SucceededUpdateWorkTodo.kt
@@ -2,4 +2,4 @@ package com.ogata_k.mobile.winp.presentation.event.work
 
 import com.ogata_k.mobile.winp.presentation.event.Event
 
-data class SucceededUpdateWorkTodo(val workId: Int, val workTodoId: Int?) : Event
+data class SucceededUpdateWorkTodo(val workId: Long, val workTodoId: Long?) : Event
