+++ app/src/main/java/com/ogata_k/mobile/winp/presentation/page/work/detail/WorkDetailUiState.kt
@@ -9,10 +9,10 @@ import java.util.Optional
 data class WorkDetailUiState(
     override val loadingState: ScreenLoadingState,
     override val basicState: BasicScreenState,
-    val workId: Int,
+    val workId: Long,
     val work: Optional<Work>,
     val inShowMoreAction: Boolean,
     val inConfirmDelete: Boolean,
     // Not nullで表示中
-    val inConfirmWorkTodoState: Int?,
+    val inConfirmWorkTodoState: Long?,
 ) : IUiState<ScreenLoadingState>
